Move chat.Object to object.Object so it can be used from different packages
