+++ chat/message.go
@@ -4,6 +4,8 @@ package chat
 
 import (
 	"log"
+
+	"github.com/joinself/self-go-sdk/pkg/object"
 )
 
 // Message represents a chat message.
@@ -15,7 +17,7 @@ type Message struct {
 	GID        string
 	ISS        string
 	Payload    map[string]interface{}
-	Objects    []*Object
+	Objects    []*object.Object
 }
 
 // NewMessage creates a chat message object.
@@ -36,7 +38,7 @@ func NewMessage(chat *Service, recipients []string, payload map[string]interface
 	if payload["objects"] != nil {
 		for _, oo := range payload["objects"].([]interface{}) {
 			o := oo.(map[string]interface{})
-			obj := NewObject(chat.fileInteractor)
+			obj := object.New(chat.fileInteractor)
 			if _, ok := o["key"]; ok {
 				err := obj.BuildFromObject(o)
 				if err != nil {
